feat: format result with prettier and unstring snapshots
BREAKING CHANGE: Your snapshots will probably need to be updated with these changes. If you don't like the changes, then take a look at the README about overriding the formatResult and disabling the un-string snapshot serializer
