+++ src/unstring-snapshot-serializer.js
@@ -1,0 +1,6 @@
+const unstringSnapshotSerializer = {
+  test: val => typeof val === 'string',
+  print: val => val,
+}
+
+export default unstringSnapshotSerializer
